SkillTrend/requirements.txt
beautifulsoup4
requests
pandas
python-dotenv
openai
ollama
nltk
gradio